Overhaul of Data Structure, CLI Revisions
DATA STRUCTURE CHANGES: Fields and Methods now abstracted to their own class. They now require a type along with the name of the object. Methods can now hold parameters. Parameters require a type and name. Method, Field and Parameter extend Formal to manage their type and name. A relationship can now be made from class1 to class2 and vice versa. CLI CHANGES: CLI updated to reflect the changes in the data structures. Also caught a bug if the user called a function, but provided no further parameters to the function.