Add enum changes from CLI to Develop (#67)
* Fixes ctrl-D bug

* Adds throws IOException

* Adds \n to error messages

* Fixes grammer in error message

* Sprint 1 Bugfixes & Enhancements

Fixed ctrl + D causing an infinite loop. It now properly exits the program. Added utility: CLI now prompts user when quitting/loading if they have unsaved changes. Also prompts when using clear to make sure they intend to delete all files.

* Added package mike

* Overhaul of Data Structure, CLI Revisions

DATA STRUCTURE CHANGES: Fields and Methods now abstracted to their own class. They now require a type along with the name of the object. Methods can now hold parameters. Parameters require a type and name. Method, Field and Parameter extend Formal to manage their type and name. A relationship can now be made from class1 to class2 and vice versa. CLI CHANGES: CLI updated to reflect the changes in the data structures. Also caught a bug if the user called a function, but provided no further parameters to the function.

* Fixed typo

"Exisit" changed to "Exist" on line 114.

* Removed unnecessary call to datastructures.java

Removed a call to datastructures on line 14 when creating a Classes object. Shouldn't be necessary, since classes.java is already imported at the top.

* More bugfixes on CLI

Added missing error messages for parameter errors, and a break statement. Also includes several missing testfiles. Build might fail on some tests, but this will be addressed in a future commit.

* Added Enums to Relationships

Relationships now manage an enum Type rather than a name. It may be an AGGREGATION, ASSOCIATION, COMPOSITION, or INHERITANCE. CLI and HelperMethods was also updated to reflect this change.

* Added missing package names

Added "mike." to the beginning of package name and import to make gradle happy.

* Added missing package names

Added "mike." to the beginning of package name and imports to make gradle happy

* Updated CLI to match changes in Develop

Updates the data structure to keep pace with the recent changes to develop. Also includes updates to tests.

* Added missing curly brace

Was missing a curly brace at the very end of the class to close the program out.

* Changed type of relationName

On line 101 relationName was changed to use a String. This ensures proper saving/loading of the json file.

* Changed Association enum to Realization

User requirements changed - they don't want associations, but realizations

* Changed Association to Realization

User requirements changed - associations now changed to realizations

Co-authored-by: phila413 <70169587+phila413@users.noreply.github.com>