Adding GUI updates to develop (#73)
* Create controller.java

Creates the controller.  It catches if the save, saveAs, load, or any of the tree functions are pressed.

* Adds class functions

When creating, renaming, or deleting classes, it will change on the gui

* Merging files from view into gui (#70)

* Create view.java

Creates the beginning view.  Also, can create class boxes in the diagram that can be dragged around.

* Adds hashmap containing all jlabels

* Update main to handle gui option

Main now checks the args array to see if user wanted to use the CLI or the GUI

* Merge view branch into gui branch (#72)

* Create view.java

Creates the beginning view.  Also, can create class boxes in the diagram that can be dragged around.

* Adds hashmap containing all jlabels

* Major refactor of View.java

Refactor of View.java to include these functionalities:
 - An 'updateClass' method that changes a class in the view in response to user actions
 - A 'deleteClass' method to remove a class from the view in response to user actions
 - An 'entityToHTML' method to properly handle creating the graphical components to a class

* Merge changes in controller to gui (#71)

* Add guiHelperMethods.java

guiHelperMethods abstracts a lot of the functions away from controller to clean up Controller.java, and make it easier to read/understand

* Major refactor of Controller.java

Controller.java had many of its functions moved to guiHelperMethods.java. As such, it is a lot easier to read/understand the class overall. 

It is still responsible for creating listeners, and calling the correct functions depending on what was clicked in the view.

* Removed old view.java

This version of view.java should have been cleared out with the merge from the view branch, but wasn't. 

Doing it here, with this commit to prevent issues.

* Update main.java

* Fixed file imports

Co-authored-by: phila413 <70169587+phila413@users.noreply.github.com>
Co-authored-by: John <54130978+JohnHynes@users.noreply.github.com>
Co-authored-by: JohnHynes <johnthehynes@gmail.com>