Major Restructure of Project Files
GUI and CLI renamed to GUIView and CLIView, and moved to a View package to more closely reflect an MVC design pattern. View and ViewInterface deleted because they served a vestigial purpose. Controller broken up, and placed into its own controller package. Classes.java renamed to Model.java to more closely reflect the MVC pattern. Resources folder for images needed in the GUI created and placed inside the GUI package.\n\nNOTE: This these files were created in the middleof the changes being made for edit mode, so some functionality of the GUI will appear broken. Once the changes for edit mode are finalized, that functionality will be restored.