File Restructure 3: Revenge of GitHub (#118)
* File Restructure 3: Revenge of GitHub

The third attempt at trying the file restructure to make way for edit mode. This was a fresh pull from develop as of ~5:30pm on 10/23, so it includes: JLine tab completion, the breakup and abstraction of Controller, the removal of the old View/ViewInterface, implementation of GUI and CLI into GUIView and CLIView in the package view, and the initial steps of edit mode - the editBox and htmlBox.

Restructuring the files to make more sense and better follow the MVC format.