Mvc restruct (#124)
* Restructures classes for mvc

MVC is now not static and passes everything through parameters.  Adds back the viewtemplate to decide if view is gui or cli.

Got rid of long empty space in the middle to remind us of a possible error

* Fixed Quitting GUI w/ Unsaved Changes

The exit listener in the gui was being created with a variable, rather than getting the value of that variable at the time the user tried to exit. Fixed.

Co-authored-by: Philip J Androwick <pjandrow@ritchie.cs.millersville.edu>
Co-authored-by: HomelessSkittle <conorsosh@gmail.com>