Cli view restruct (#130)
* CLI Change to Enable Testing

Abstracted the REPL loop of the cli to two different functions. One now begins the REPL loop, the other evaluates the command created during the REPL loop. This enables us to actually test the CLI, imagine that.

* Added a sudo option to CLI

Added an undocumented sudo option for quit, clear, and load to allow the user to force those changes to happen without going through the prompt for them. Also *conveniently* makes it easier to test these functions later.

* Fixed wonky indent

Co-authored-by: SerbianSensation <stefangligorevic@gmail.com>