Add the Sprint 3 changes to Usage Branch (#134)
* Continuous integration testing (#93)

* Added initial travis yml file for continuous integration testing

* Changed build command and added test command

* Added all OS

* Continuous integration testing (#95)

* Added initial travis yml file for continuous integration testing

* Changed build command and added test command

* Added all OS

* Renamed travis to gradle, added 'on' trigger event

* Added jobs, specified windows as OS to run on

* Removed build line under test job

* Added push to 'on' trigger event

* Changed push from 'on' trigger event to pull_request

* Removed old travis.yml file that is no longer needed (#96)

* Relationship GUI (#91)

* Added GUIRelationship.java
- Does the math to draw a relationship between two labels

* Implemented Line and refactored GUIRelationship

* Draws Relationships into GUI (#97)

* Added GUIRelationship.java
- Does the math to draw a relationship between two labels

* Implemented Line and refactored GUIRelationship

* Adds Relationship Line to GUI

* Cleaned Up Line Changes

Removed some validates/repaints because they didn't do anything. Removed useless comments. Moved action listeners from GUI.java to Controller.java.

* Refactor of GUIRelationship.java (#98)

Refactor GUIRelationship methods to minimize math and improve efficiency.

* Relationship GUI Updates (#99)

* Added GUIRelationship.java
- Does the math to draw a relationship between two labels

* Implemented Line and refactored GUIRelationship

* Adds Relationship Line to GUI

* Cleaned Up Line Changes

Removed some validates/repaints because they didn't do anything. Removed useless comments. Moved action listeners from GUI.java to Controller.java.

* Enhanced GUIRelationship.java

Refactored getEdgeIntersectionPoint method to use much less math, and (hopefully) feature improved performance. Caused changes to other files to facilitate these changes. Added getters/setters to View and Controller to remove warnings.

* Implemented stylized line support

* Update src/main/java/mike/gui/GUI.java

Fixed a typo I made with a manual fix

* Fixed relationship resize bug
Refatored code into an "update" method in Line
 - Can simply put JLabels in as parameters now instead of points
Made Line fields private

* Refactored and move some relationship code
Added in dashed line for Realization type
Implemented working loading for Relationships

* Updated UID

Line class now has a properly generated UID

* Removed useless file and redundant code

* Replaced Math.pow function calls and moved some code

* Privated some functions in Line

* Self-Relationships (#100)

Stylized lines for creating a recursive relationship (from one class to itself).

* JLine stuff (#115)

Added tab completion feature to the console interface.

* Tab Completion Improvements (#116)

Add enhancement to tab completion in CLI where completers only offer applicable command arguments.

* File Restructure 3: Revenge of GitHub (#118)

* File Restructure 3: Revenge of GitHub

The third attempt at trying the file restructure to make way for edit mode. This was a fresh pull from develop as of ~5:30pm on 10/23, so it includes: JLine tab completion, the breakup and abstraction of Controller, the removal of the old View/ViewInterface, implementation of GUI and CLI into GUIView and CLIView in the package view, and the initial steps of edit mode - the editBox and htmlBox.

Restructuring the files to make more sense and better follow the MVC format.

* Adds edit mode (#120)

* Adds edit mode

Does all the functionality of edit mode except for relationships

* Revert to CLIView

An older version of the CLIView was being uploaded here - changed to the be current version of develop.

* Merge EditModeRelationships into Develop (#122)

* Adds add/delete button for relationships

Adds an add and a delete relationship button when in an edit mode for a class.  It still has some bugs though.

* Edit Mode Features and Bughunts

Save/SaveAs/Load Buttons now get disabled when in Edit Mode to prevent user from doing janky things while editing. Edit Mode and Add Class buttons get disabled while actively editing a class. The actively edited class now shows up on top of all other things in the GUI. The load function now properly loads and shows relationships.

* Mvc restruct (#124)

* Restructures classes for mvc

MVC is now not static and passes everything through parameters.  Adds back the viewtemplate to decide if view is gui or cli.

Got rid of long empty space in the middle to remind us of a possible error

* Fixed Quitting GUI w/ Unsaved Changes

The exit listener in the gui was being created with a variable, rather than getting the value of that variable at the time the user tried to exit. Fixed.

* update README (#127)

Readme updated for sprint 3

* Added tab completion for relationship types (#128)

* CLI Change to Enable Testing (#129)

Abstracted the REPL loop of the cli to two different functions. One now begins the REPL loop, the other evaluates the command created during the REPL loop. This enables us to actually test the CLI, imagine that.

The program no longer runs in the constructor.

Co-authored-by: HomelessSkittle <conorsosh@gmail.com>

* Cli view restruct (#130)

* CLI Change to Enable Testing

Abstracted the REPL loop of the cli to two different functions. One now begins the REPL loop, the other evaluates the command created during the REPL loop. This enables us to actually test the CLI, imagine that.

* Added a sudo option to CLI

Added an undocumented sudo option for quit, clear, and load to allow the user to force those changes to happen without going through the prompt for them. Also *conveniently* makes it easier to test these functions later.

* Fixed wonky indent

Co-authored-by: SerbianSensation <stefangligorevic@gmail.com>

* Updates CLIView (#131)

Found a small bug when creating tests.  This changes commands[11] to commandUsage[11] on line 270.  When a user types in the wrong amount of arguments in "delete parameter", line 270 is printed, and we don't want to print out the user's 12th argument!

Co-authored-by: SerbianSensation <56570006+SerbianSensation@users.noreply.github.com>
Co-authored-by: SerbianSensation <stefangligorevic@gmail.com>
Co-authored-by: JohnHynes <johnthehynes@gmail.com>
Co-authored-by: John <54130978+JohnHynes@users.noreply.github.com>
Co-authored-by: phila413 <70169587+phila413@users.noreply.github.com>
Co-authored-by: Philip J Androwick <pjandrow@ritchie.cs.millersville.edu>
Co-authored-by: HomelessSkittle <conorsosh@gmail.com>
Co-authored-by: Conor S <70168298+HomelessSkittle@users.noreply.github.com>
Co-authored-by: Will Killian <william.killian@millersville.edu>
Co-authored-by: Will Killian <william.killian@outlook.com>