Added Visibility Creation to Model (#143)
* Added Visibility Creation to Model

This commit focuses primarily on changing the Model, Entity, Field and Method files to allow for the creation of a visibility variable for Methods and Fields. Following this, the tests, the CLI, and some GUI components needed to be updated to handle the new functionality. NOTE: In this version, edit mode had to be disabled to allow the .jar file to compile, since some controllers in the GUI needed functions to edit visibility, which would be impossible prior to this.

* Added null check to visibility.

CLIView now checks for an improper visibility argument being passed by the user. Also changed the name of the visibility checking function in Entity.