Adds functionality to edit Visibility (#144)
* Adds functionality to edit Visibility

Adds the ability for the user to edit the visibility of a field or method in both the CLI and in the GUI. This branch restores full functionality to the GUI that was broken with the Add Visibility PR.

* Update ModelTest

Updated ModelTest to include tests for editing visibility. Entity checks input visibility to make sure it is valid.

* Hopefully Improved ModelTests

Changed how we do some testing to hopefully improve code coverage enough that the numbers go up.

* More Updates to ModelTest.java

Removed unnecessary equality test in Model.empty(); Added more scope to testGetRelationship. Removed unused boolean from LoadController.