Added file chooser to save and load controller
Changed the saveWithInput method in save controller, and the load controller to use a JFileChooser for graphical input of choosing where to save a file. Works as you'd expect, but it says open instead of save or load in the file chooser, and I'm not sure how to change that.