Fixed Nondestructive Loading
Phil found an issue when loading that the previous things in the GUI weren't getting cleared out, so when loading a new set of data, it would all just get added to the old stuff. Loading now properly clears out the pane and removes all information from the previous file.